requirements.txt
streamlit
crewai
openai
langchain-openai
python-dotenv
pandas